pyproject.toml
[project]
name = "prox-lora"
version = "0.1.0"
description = "Karol Chojnacki MSc Thesis - playground"
readme = "README.md"
authors = [{ name = "Karol Chojnacki" }, { name = "Marcin Wrochna" }]
urls.repository = "https://github.com/VeryDisappointedSalad/Prox-LoRA"
requires-python = ">=3.13,<3.14"
dependencies = [
    "clearml>=2.1.3",
    "einops",
    "huggingface-hub",
    "ipykernel",
    "ipywidgets",
    "jupyter>=1.1",
    "lightning>=2.6.1",
    "matplotlib>=3.10",
    "numpy>=2.2",
    "open_clip_torch",
    # "opencv-python>=4.12",
    "pandas>=2.3",
    "pillow>=11.0.0",
    "psutil",              # logging CPU stats.
    "pydantic>=2.12",
    "python-dotenv",
    "rich",
    "ruamel.yaml==0.18.*",
    "scikit-learn>=1.7",
    "scipy",
    "seaborn",
    "tensorboardX>=2.6",
    "timm>=1.0",
    "transformers>=4.55",
    "tqdm",
    "typing-extensions",
    "foolbox>=3.3.4",
]


# Published optional dependencies, or "extras". We use them for different architectures (CPU/CUDA/Apple Silicon/...).
# See https://docs.astral.sh/uv/guides/integration/pytorch/#configuring-accelerators-with-environment-markers
[project.optional-dependencies]
cpu = ["torch>=2.10", "torchvision>=0.25"]
cu128 = ["torch>=2.10", "torchvision>=0.25", "nvidia-ml-py"]
cu130 = ["torch>=2.10", "torchvision>=0.25", "nvidia-ml-py"]

# Local dependencies for development.
[dependency-groups]
dev = [
    "mypy",
    "ruff",
    "pandas-stubs",
    "types-pyyaml>=6.0.12",
    "types-toml",
    "types-tqdm>=4.67",
    "scipy-stubs>=1.16",
]


[tool.uv]
conflicts = [[{ extra = "cpu" }, { extra = "cu128" }, { extra = "cu130" }]]  # Can't have both CPU and CUDA versions installed.


[tool.uv.sources]
torch =       [{extra="cpu", index="pytorch-cpu"}, {extra="cu128", index="pytorch-cu128"}, {extra="cu130", index="pytorch-cu130"}]
torchvision = [{extra="cpu", index="pytorch-cpu"}, {extra="cu128", index="pytorch-cu128"}, {extra="cu130", index="pytorch-cu130"}]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true


[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"


################################### Mypy ######################################
[tool.mypy]
strict = true
disallow_untyped_calls = true
untyped_calls_exclude = ["mlflow", "transformers"]
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_column_numbers = true
strict_equality = true
strict_concatenate = true
strict_bytes = true
implicit_reexport = false
exclude_gitignore = true
exclude = ["tmp/"]
follow_untyped_imports = false

# Libraries we know don't have types (even as separate packages).
# In particular, HuggingFace doesn't like types:
#     https://github.com/huggingface/transformers/pull/18485
[[tool.mypy.overrides]]
module = [
    "transformers.*",
    "datasets.*",
    "accelerate.*",
    "trl.*",
    "tokenizers.*",
    "scipy.*",
    "mlflow.*",
    "joblib",
    "numba",
    "open_clip.*",
    "open_clip_torch",
]
ignore_missing_imports = true
implicit_reexport = true
check_untyped_defs = false


# Libraries that mostly have types but aren't marked as such with 'py.typed'.
# E.g.: https://github.com/google-deepmind/optax/issues/1388
[[tool.mypy.overrides]]
module = [
    "timm.*",
    "torchvision.*",
    "optax.*",
    "orbax.*",
    "etils.*",
    "pynvml",
    "fsspec.*",
    "numpydantic",
    "augmax",
    "ml_collections",
    "msgpack",
    "tree",
]
follow_untyped_imports = true
check_untyped_defs = true
implicit_reexport = true


################################### Ruff ######################################
[tool.ruff]
line-length = 120
target-version = "py313"    # Minimum supported version.
src = ["src", "tests"]
extend-exclude = ["tmp/**"]

[tool.ruff.lint]
# See https://docs.astral.sh/ruff/rules/
select = [
    "ASYNC",
    "B",
    "C4",
    "COM818",
    "DTZ",
    "E",
    "EXE",
    "F",
    "FA",
    "FBT",
    "FURB",
    "I",
    "ICN",
    "INP",
    "ISC",
    "LOG",
    "NPY",
    "PD",
    "PERF",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "Q",
    "RUF",
    "S102",
    "S202",
    "S306",
    "S307",
    "SLF",
    "SLOT",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "F722",    # 'forward-annotation-syntax-error' - may conflict with array typing.
    "T201",    # We use print statements.
    "B009",    # 'Do not call `getattr` with a constant attribute' - false positive when called with `default=`.
    "C408",    # 'Unnecessary dict call (rewrite as a literal)' - sometimes `dict(a=b, )` is more readable and shorter.
    "C420",    # 'Unnecessary dict comprehension, use dict.fromkeys instead' - dict comprehension is more readable.
    "E501",    # 'Line too long' - we let ruff-format handle that, allowing long docstrings and comments in particular.
    "E731",    # 'Do not assign lambda, use def' – lambda can be definitely more readable.
    "ISC003",  # 'Explicitly concatenated string should be implicitly concatenated' - we prefer explicit concatenation.
    "PD010",   # '.pivot_table is preferred to .pivot and .unstack' - unstack can be way more readable, pivot_table can do accidental agg.
    "PLC0414", # 'Useless from x import y as y' - this used use to by mypy to explicitly re-export something (see `--no-implicit-reexport`).
    "PLR0402", # Idem.
    "PLR0912", # 'Too many branches' - well, sometimes you need to.
    "PLR0911", # 'Too many return statements' - idem.
    "PLR0913", # 'Too many arguments' - idem.
    "PLR0915", # 'Too many statements' - idem.
    "PLR1714", # 'Consider merging multiple comparisons: `x in (a, b)`' - sometimes less readable.
    "PLR2004", # 'Magic value used in comparison' - usually false positives.
    "PLR5501", # 'Use `elif` instead of `else` then `if`' - sometimes less readable (another case vs an unrelated check).
    "PLW2901", # 'for loop variable overwritten' - I don't think it's risky when it's a normal assignment with '='.
    "PT018",   # 'Assertion should be broken down into multiple parts' - annoying when you have the same message for all.
    "PTH123",  #  'open() should be replaced by Path.open()' - Path.open() is less idiomatic.
    "PYI036",  # 'bad-exit-annotation' - __exit__ types are very verbose and almost never salient.
    "PYI041",  # 'Use `float` instead of `int | float`' - the latter may be informative.
    "RUF001",  # 'String contains ambiguous character' - we use UTF-8 characters like multiplication "×".
    "RUF002",  # 'Docstring contains ambiguous character' - idem.
    "RUF003",  # 'Comment contains ambiguous character' - idem.
    "RUF009",  # function-call-in-dataclass-defaults - too many false positives for frozen dataclasses with mutable members.
    "RUF100",  # 'Unused noqa'.
]
unfixable = [
    "B905", # Fix defaults to zip(strict=False), which is not what we want.
]

[tool.ruff.lint.isort]
known-first-party = ["prox_lora"]
known-third-party = ["datasets", "wandb"]
split-on-trailing-comma = false

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["pathlib.PurePath"]

[tool.ruff.format]
skip-magic-trailing-comma = true
